requirements.txt
python-telegram-bot[job-queue]==20.8
gspread==5.12.4
google-auth==2.25.2
requests==2.31.0
pytz==2023.3
python-dotenv==1.0.0
shapely==2.0.2
aiohttp==3.9.1
python-dateutil>=2.9
rapidfuzz>=3.6

# Optional admin tools (install with: pip install psutil rich)
# psutil==5.9.5
# rich==13.5.2